Move all substrait functions into a dedicated file to reduce cfg annotation proliferation
