+++ rust/lance-datafusion/src/lib.rs
@@ -9,4 +9,6 @@ pub mod logical_expr;
 pub mod planner;
 pub mod projection;
 pub mod sql;
+#[cfg(feature = "substrait")]
+pub mod substrait;
 pub mod utils;
